+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Compile time check for padding now works reliably
 - Add `no_ctors` attribute which allows the user to write custom constructors.
 
+### Added
+
+- Allow field types which also derive `Mmio` as long as they are annotated with
+  `#[mmio(inner)]`
+
 ## [v0.3.0] - 2025-02-26
 
 ### Changed